Cleanup settings vm, move functionality in to Help
